# openapi.yaml
openapi: 3.0.3
info:
  title: API de avistamientos de aves
  version: 1.0.0
  description: |
    API local del taller de ISW-2. Los datos persisten en SQLite.
    POST y PUT reciben los cuatro campos editables; el id lo asigna el sistema.
    PUT reemplaza esos campos y no crea recursos inexistentes.
    Los textos se recortan en sus extremos y no pueden quedar vacíos.
    Los cuerpos JSON tienen un límite de 64 KiB.
servers:
  - url: http://127.0.0.1:5000
tags:
  - name: Avistamientos
paths:
  /avistamientos:
    get:
      tags: [Avistamientos]
      operationId: listarAvistamientos
      summary: Listar todos los avistamientos
      responses:
        '200':
          description: Lista ordenada por id; puede estar vacía.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Avistamiento'
        '500':
          $ref: '#/components/responses/ErrorInterno'
    post:
      tags: [Avistamientos]
      operationId: crearAvistamiento
      summary: Crear un avistamiento
      requestBody:
        $ref: '#/components/requestBodies/DatosAvistamiento'
      responses:
        '201':
          description: Avistamiento creado con id generado.
          headers:
            Location:
              description: Ruta relativa del nuevo recurso.
              schema:
                type: string
                example: /avistamientos/1
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Avistamiento'
        '400':
          $ref: '#/components/responses/DatosInvalidos'
        '413':
          $ref: '#/components/responses/CuerpoExcesivo'
        '500':
          $ref: '#/components/responses/ErrorInterno'
  /avistamientos/resumen:
    get:
      tags: [Avistamientos]
      operationId: resumirAvistamientos
      summary: Contar los registros por especie
      description: Agrupa por texto exacto de especie; distingue mayúsculas y tildes.
      responses:
        '200':
          description: Lista de conteos ordenada por especie; puede estar vacía.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: false
                  required: [especie, cantidad]
                  properties:
                    especie:
                      type: string
                    cantidad:
                      type: integer
                      minimum: 1
              example:
                - especie: colibrí
                  cantidad: 2
        '500':
          $ref: '#/components/responses/ErrorInterno'
  /avistamientos/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: Identificador asignado por SQLite. Fuera del rango válido responde 404.
        schema:
          type: integer
          format: int64
          minimum: 1
          maximum: 9223372036854775807
    get:
      tags: [Avistamientos]
      operationId: obtenerAvistamiento
      summary: Consultar un avistamiento
      responses:
        '200':
          description: Registro encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Avistamiento'
        '404':
          $ref: '#/components/responses/NoEncontrado'
        '500':
          $ref: '#/components/responses/ErrorInterno'
    put:
      tags: [Avistamientos]
      operationId: reemplazarAvistamiento
      summary: Reemplazar los cuatro campos editables de un avistamiento
      description: Conserva el id. Si el recurso no existe, devuelve 404 antes de validar el cuerpo.
      requestBody:
        $ref: '#/components/requestBodies/DatosAvistamiento'
      responses:
        '200':
          description: Registro actualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Avistamiento'
        '400':
          $ref: '#/components/responses/DatosInvalidos'
        '404':
          $ref: '#/components/responses/NoEncontrado'
        '413':
          $ref: '#/components/responses/CuerpoExcesivo'
        '500':
          $ref: '#/components/responses/ErrorInterno'
    delete:
      tags: [Avistamientos]
      operationId: eliminarAvistamiento
      summary: Eliminar un avistamiento
      responses:
        '200':
          description: Eliminación confirmada con un cuerpo JSON.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [mensaje, id]
                properties:
                  mensaje:
                    type: string
                    example: Avistamiento eliminado.
                  id:
                    type: integer
                    format: int64
                    minimum: 1
        '404':
          $ref: '#/components/responses/NoEncontrado'
        '500':
          $ref: '#/components/responses/ErrorInterno'
components:
  requestBodies:
    DatosAvistamiento:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DatosAvistamiento'
          example:
            especie: colibrí
            lugar: Jardín Botánico de Bogotá
            fecha: '2026-09-29'
            observador: Nicolás
  schemas:
    TextoObligatorio:
      type: string
      minLength: 1
      pattern: '\S'
      description: Texto con al menos un carácter no blanco; se recortan los espacios exteriores.
    Fecha:
      type: string
      format: date
      pattern: '^[0-9]{4}-[0-9]{2}-[0-9]{2}$'
      description: Fecha real del calendario entre los años 0001 y 9999.
      example: '2026-09-29'
    DatosAvistamiento:
      type: object
      additionalProperties: false
      required: [especie, lugar, fecha, observador]
      properties:
        especie:
          $ref: '#/components/schemas/TextoObligatorio'
        lugar:
          $ref: '#/components/schemas/TextoObligatorio'
        fecha:
          $ref: '#/components/schemas/Fecha'
        observador:
          $ref: '#/components/schemas/TextoObligatorio'
    Avistamiento:
      type: object
      additionalProperties: false
      required: [id, especie, lugar, fecha, observador]
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
          maximum: 9223372036854775807
          readOnly: true
        especie:
          $ref: '#/components/schemas/TextoObligatorio'
        lugar:
          $ref: '#/components/schemas/TextoObligatorio'
        fecha:
          $ref: '#/components/schemas/Fecha'
        observador:
          $ref: '#/components/schemas/TextoObligatorio'
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string
  responses:
    DatosInvalidos:
      description: Campos ausentes o inválidos, JSON mal formado, cuerpo no objeto o Content-Type incorrecto.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: 'Faltan campos obligatorios: lugar, fecha, observador.'
    NoEncontrado:
      description: El avistamiento no existe.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Avistamiento no encontrado.
    CuerpoExcesivo:
      description: El cuerpo supera 64 KiB.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ErrorInterno:
      description: Error interno del servidor.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Error interno del servidor.